Use "help" cursor instead of "pointer"
I believe this is a more appropriate cursor as for tooltips. I read about the Safari bug in the comment, and I wasn't able to reproduce it on my iOS 13 / iPhone device nor my Macbook. It might be it's gone!

Please let me know what do you think. Keep up the good work!